requirements.txt
flask = "*"
flask-sqlalchemy = "*"
flask-marshmallow = "*"
marshmallow-sqlalchemy = "*"
flask-restful = "*"
flask-jwt-extended = "*"
flask-cors = "*"
flask-migrate = "*"
